1:级联选择框调整  2:推送路径过滤空字符串

+++ v5configure_server/src/main/java/com/yunwen/android/v5configure_s/dao/CategoryDAO.java
@@ -5,7 +5,7 @@ import com.yunwen.android.v5configure_s.pojo.DeviceGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CategoryDAO extends JpaRepository<Category,Integer> {
+public interface CategoryDAO extends JpaRepository<Category, Long> {
 
 //    List<DeviceGroup> findByName(String deviceName);
