+++ v5configure_server/src/main/java/com/yunwen/android/v5configure_s/service/CategoryService.java
@@ -29,7 +29,7 @@ public class CategoryService {
 //
 //    }
 
-    public void removeCategoryById(Integer id) {
+    public void removeCategoryById(Long id) {
         categoryDAO.deleteById(id);
     }
