+++ v5configure_server/src/main/java/com/yunwen/android/v5configure_s/service/DeviceGroupService.java
@@ -27,7 +27,7 @@ public class DeviceGroupService {
 //
 //    }
 
-    public void removeDeviceGroupById(Integer id) {
+    public void removeDeviceGroupById(Long id) {
         deviceGroupDAO.deleteById(id);
     }
