+++ v5configure_server/src/main/java/com/yunwen/android/v5configure_s/service/PushContentInfoService.java
@@ -29,7 +29,7 @@ public class PushContentInfoService {
 //
 //    }
 
-    public void removePushContentById(Integer id) {
+    public void removePushContentById(Long id) {
         pushContentDAO.deleteById(id);
     }
